+++ src/MetaheuristicsPlatform/Algorithms/PSO/Scientific/FullyInformedPsoReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.PSO.Scientific;
+
+public static class FullyInformedPsoReferences
+{
+    public static ScientificReference MendesKennedyNeves2004 { get; } =
+        new(
+            "Mendes, Kennedy & Neves",
+            2004,
+            "The Fully Informed Particle Swarm: Simpler, Maybe Better",
+            "IEEE Transactions on Evolutionary Computation 8(3), 204-210",
+            "10.1109/TEVC.2004.826074");
+}